Split static config
