+++ backends-velox/src/test/scala/org/apache/gluten/execution/VeloxRoughCostModelSuite.scala
@@ -16,7 +16,8 @@
  */
 package org.apache.gluten.execution
 
-import org.apache.gluten.config.GlutenConfig
+import org.apache.gluten.config.GlutenConfig._
+import org.apache.gluten.config.GlutenStaticConfig._
 
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.execution.ProjectExec
@@ -41,12 +42,12 @@ class VeloxRoughCostModelSuite extends VeloxWholeStageTransformerSuite {
   }
 
   override protected def sparkConf: SparkConf = super.sparkConf
-    .set(GlutenConfig.RAS_ENABLED.key, "true")
-    .set(GlutenConfig.RAS_COST_MODEL.key, "rough")
-    .set(GlutenConfig.VANILLA_VECTORIZED_READERS_ENABLED.key, "false")
+    .set(RAS_ENABLED.key, "true")
+    .set(RAS_COST_MODEL.key, "rough")
+    .set(VANILLA_VECTORIZED_READERS_ENABLED.key, "false")
 
   test("fallback trivial project if its neighbor nodes fell back") {
-    withSQLConf(GlutenConfig.COLUMNAR_FILESCAN_ENABLED.key -> "false") {
+    withSQLConf(COLUMNAR_FILESCAN_ENABLED.key -> "false") {
       runQueryAndCompare("select c1 as c3 from tmp1") {
         checkSparkOperatorMatch[ProjectExec]
       }
@@ -54,7 +55,7 @@ class VeloxRoughCostModelSuite extends VeloxWholeStageTransformerSuite {
   }
 
   test("avoid adding r2c whose schema contains complex data types") {
-    withSQLConf(GlutenConfig.COLUMNAR_FILESCAN_ENABLED.key -> "false") {
+    withSQLConf(COLUMNAR_FILESCAN_ENABLED.key -> "false") {
       runQueryAndCompare("select array_contains(c3, 0) as list from tmp1") {
         checkSparkOperatorMatch[ProjectExec]
       }
